Adds notes management and internal vim editor
Notes can now be managed with the Shift+T Keybinding, I will change this to Shift+N in a later commit, restoring translation features with Shift+T (currently not accessible)

Internal editor added now with Ctrl+X (for messages) and as default for notes management. This has simple vim keybindings including navigation, insert and visual mode, and modifiers like dd (to delete a line). It is minimal, and very incomplete, but it works. Use :wq to send messages / save a note.

There were also minor imporvements and changes to the bot - but these are just being staged and aren't useful atm.